rename projects styles folder
