+++ components/projects-components/MediaCard.tsx
@@ -5,7 +5,7 @@ import Typography from '@mui/material/Typography';
 import { Grid } from '@mui/material';
 import Link from 'next/link';
 import { motion } from "framer-motion";
-import styles from "../../styles/projects-styles/Card.module.css"
+import styles from "../../styles/projects-components/Card.module.css"
 
 export default function MediaCard({ title, link, src }) {
     const boxHover = {
